Add static data system, reorganize bin scripts
Instead of storing some EVE-related data (like a list of skills
and their IDs/names) in a DB, be lazy and store them in memory!
