+++ back-end/bin/skills/dump-skills.js
@@ -1,0 +1,17 @@
+/**
+ * Dumps a map of all skills in the game. Maps skill IDs to skill data.
+ * 
+ * Usage:
+ * $ node bin/dump-skills.js > path/to/target.json
+ */
+
+const fetchSkills = require('./fetch-skills');
+
+fetchSkills()
+.then(function([skillsMap, groupsMap]) {
+  console.log('// Generated by bin/skills/dump-skills.js');
+  console.log(JSON.stringify(skillsMap, null, 2));
+})
+.catch (function(err) {
+  console.error('ERROR:', err);
+});
