+++ back-end/bin/skills/generate-fake-skills.js
@@ -1,0 +1,72 @@
+/**
+ * Generates a set of skills for a character. Assigns levels randomly.
+ * 
+ * Usage:
+ * $ node bin/generate-fake-skills.js > path/to/target.json
+ */
+const fetchSkills = require('./fetch-skills');
+
+
+const GROUP_DISPLAY_ORDER = [
+  257,    // Spaceship Command
+  275,    // Navigation
+  1216,   // Engineering
+  1240,   // Subsystems
+  1210,   // Armor
+  1209,   // Shields
+
+  1213,   // Targeting
+  255,    // Gunnery
+  256,    // Missiles
+  273,    // Drones
+  272,    // Electronic Systems
+  1217,   // Scanning
+
+  269,    // Rigging
+  278,    // Social
+  258,    // Fleet Support
+  266,    // Coporation Management
+  274,    // Trade
+  1220,   // Neural Enhancement
+
+  268,    // Production
+  270,    // Science
+  1218,   // Resource Processing
+  1241,   // Planet Management
+  1545,   // Structure Management
+];
+
+fetchSkills()
+.then(function([skillsMap, groupsMap]) {
+  let out = [];
+  for (let i = 0; i < GROUP_DISPLAY_ORDER.length; i++) {
+    let group = groupsMap[GROUP_DISPLAY_ORDER[i]];
+    out.push(group);
+
+    // Sort skills alphabetically
+    group.skills.sort((a, b) => {
+      return a.name.localeCompare(b.name);
+    });
+
+    // Assign levels randomly
+    for (let j = 0; j < group.skills.length; j++) {
+      let skill = group.skills[j];
+      delete skill.description;
+      delete skill.rank;
+      delete skill.primaryAttribute;
+      delete skill.secondaryAttribute;
+      delete skill.groupId;
+      if (Math.random() < 0.5) {
+        skill.level = -1;
+      } else {
+        skill.level = Math.ceil(Math.random() * 5);
+      }
+    }
+  }
+
+  console.log('// Generated by bin/skills/generate-fake-skills.js');
+  console.log(JSON.stringify(out, null, 2));
+})
+.catch (function(err) {
+  console.error('ERROR:', err);
+});
